Analysis of US hospital data
'best' reads the outcome-of-care-measures.csv ﬁle and returns the name of the hospital that has the best (i.e. lowest) 30-day mortality for the speciﬁed outcome in that state.
'rankhospital' reads the outcome-of-care-measures.csv ﬁle and returns the name of the hospital that has the ranking speciﬁed by the num argument
'rankall' returns a 2-column data frame containing the hospital in each state that has the ranking speciﬁed in num.